Optional webRequest/webRequestBlocking permissions
Make webRequest and webRequestBlocking optional.

The permissions are now requested on demand when a request with "Clean
Request" mode is sent.

This is especially useful for scenarios where extensions with webRequest
permissions are blocked, which is possible in Chrome with policies:
https://chromeenterprise.google/policies/#ExtensionSettings.

Browsers don't seem to ask the user when those permissions are
requested. So this should not result in a worse user experience.