Silent webRequest permission request and fallback
Requesting webRequest permissions only for "Clean Request" mode does not
work because RESTer relies on them to set the Cookie header.

Instead this makes RESTer silently request webRequest permissions on the
first request (it seems necessary to request permissions only on user
interaction). If it fails, RESTer logs a warning to the console but
proceeds.

In the future a warning in the UI might be useful.