Support formatting NaN in JSON
NaN seems to be a common extension to JSON. Probably because JavaScript
supports it. RESTer doesn't do anything with JSON except making it
easier to read. It seems reasonable to try to support JSON extensions as
long as they don't make the code harder to maintain. This adds support
for .